+++ 3K.Application/Features/SandikIslemleri/DTOs/SandikIcerikDto.cs
@@ -7,9 +7,9 @@ namespace _3K.Application.Features.SandikIslemleri.DTOs
         public string? OlcuResmiPozNo { get; set; }
         public string BarkodNo { get; set; } = string.Empty;
         public string Aciklama { get; set; } = string.Empty;
-        public int IstenenAdet { get; set; }
-        public int KonulanAdet { get; set; }
-        public int EksikAdet { get; set; }
+        public decimal IstenenAdet { get; set; }
+        public decimal KonulanAdet { get; set; }
+        public decimal EksikAdet { get; set; }
         public int DurumId { get; set; }
         public string DurumMetni { get; set; } = string.Empty;
         public string? PaketleyenBasHarf { get; set; }
@@ -19,20 +19,20 @@ namespace _3K.Application.Features.SandikIslemleri.DTOs
 
         // Saha/Yedek manuel malzeme alanları
         public string? Isim { get; set; }
-        public int Miktar { get; set; }
+        public decimal Miktar { get; set; }
         public int? BirimId { get; set; }
         public string? BirimMetni { get; set; }
 
         // Parçalı karşılama (Madde 2)
-        public int StokKarsilanan { get; set; }
-        public int ProjeKarsilanan { get; set; }
-        public int TedarikciKarsilanan { get; set; }
+        public decimal StokKarsilanan { get; set; }
+        public decimal ProjeKarsilanan { get; set; }
+        public decimal TedarikciKarsilanan { get; set; }
 
         // Kaynak proje bilgisi (Saha/Yedek sandıklara projeden eklenen ürünler için)
         public string? KaynakProjeNo { get; set; }
 
         // Backend-hesaplanan alanlar (Dumb UI — KURAL 3)
-        public int KalanMiktar { get; set; }
+        public decimal KalanMiktar { get; set; }
         public int GenelDurumId { get; set; }
         public string GenelDurumMetni { get; set; } = string.Empty;
     }
